# Conflicts:
#	obm_server/obm/model/managed_battle_map.py
#	obm_server/obm/routers/image_data.py

+++ obm_server/obm/routers/battle_map.py
@@ -25,7 +25,7 @@ class BattleMapInfo(BaseModel):
     revision: int
     token_action_count: int
 
-    background_pixels_per_meter: int
+    background_pixels_per_meter: float
     background_media_type: Optional[str]
